requirements.txt
Flask==3.1.0
ollama==0.4.7
requests==2.32.3
Flask-Session==0.8.0
python-dotenv==1.0.1